chore: Remove libbpf and darwin build tags
