+++ .github/workflows/c.yml
@@ -21,7 +21,16 @@ jobs:
     steps:
       - uses: actions/checkout@v4.1.1
       - name: Install libbpf
+        working-directory: bpfassets/libbpf
         run: |
-          sudo apt-get update && sudo apt-get install -y libelf-dev libbpf-dev
+          sudo apt remove libbpf-dev
+          sudo apt install libelf-dev
+          mkdir temp-libbpf
+          cd temp-libbpf
+          git clone https://github.com/libbpf/libbpf
+          cd libbpf/src
+          sudo make install_headers
+          sudo make install_uapi_headers
+          sudo prefix=/usr  BUILD_STATIC_ONLY=y make install
       - name: Build
         run: make -C bpfassets/libbpf
