+++ e2e/platform-validation/platform_validation_metric_test.go
@@ -1,6 +1,3 @@
-//go:build bcc || libbpf
-// +build bcc libbpf
-
 /*
 Copyright 2023.
