+++ pkg/manager/manager_test.go
@@ -1,6 +1,3 @@
-//go:build !darwin
-// +build !darwin
-
 package manager
 
 import (
