--- pkg/utils/util_darwin.go
@@ -1,27 +1,0 @@
-//go:build darwin && !linux
-// +build darwin,!linux
-
-/*
-Copyright 2021.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package utils
-
-import "encoding/binary"
-
-func GetCgroupIDFromPath(byteOrder binary.ByteOrder, path string) (uint64, error) {
-	// to do for Mac OS how to get CGROUP ID from path
-	return uint64(0), nil
-}
